backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
django-cors-headers
djangorestframework-stubs==3.16.8
psycopg2-binary
Pillow
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
